+++ frontend/src/Pioreactors.jsx
@@ -1099,7 +1099,7 @@ function SettingsActionsDialog({ unit, experiment, jobs, setLabel, label, disabl
                       size="small"
                       variant="outlined"
                       label="XR"
-                      sx={{ ml: 0.5, height: 18, fontSize: "0.65rem" }}
+                      sx={{ ml: 0.5, height: 18, fontSize: "0.65rem", "& .MuiChip-label": { px: "5px" } }}
                     />
                   ) : null}
                 </Typography>
@@ -2627,7 +2627,7 @@ function PioreactorCard({unit, isUnitActive, experiment, config, originalLabel, 
                     size="small"
                     variant="outlined"
                     label="XR"
-                    sx={{ ml: 0.5, height: 18, fontSize: "0.60rem" }}
+                    sx={{ ml: 0.5, height: 18, fontSize: "0.65rem", "& .MuiChip-label": { px: "5px" } }}
                   />
                 ) : null}
               </Typography>
@@ -2674,7 +2674,7 @@ function PioreactorCard({unit, isUnitActive, experiment, config, originalLabel, 
                           size="small"
                           variant="outlined"
                           label="XR"
-                          sx={{ ml: 0.5, height: 18, fontSize: "0.65rem" }}
+                          sx={{ ml: 0.5, height: 18, fontSize: "0.65rem", "& .MuiChip-label": { px: "5px" } }}
                         />
                       ) : null}
                     </Typography>